backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

# Database drivers
neo4j

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Data processing
pyyaml
python-dateutil

# Utility
httpx
aiofiles

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
